WIP added a new suite for bootstrap restart, fixed imports and renamed fixtures.py to conftest.py
