+++ tests/late_nodes/test_delayed.py
@@ -1,14 +1,12 @@
 import time
 from pytest_testconfig import config as testconfig
-from kubernetes import client
 
-from tests.deployment import create_deployment, delete_deployment
-from tests.fixtures import set_namespace, load_config, init_session, set_docker_images, session_id, DeploymentInfo, init_session
+from tests.deployment import create_deployment
+from tests.conftest import DeploymentInfo
 from tests.test_bs import setup_clients, save_log_on_exit, setup_bootstrap, create_configmap, add_curl, start_poet
-from tests.test_bs import current_index, wait_genesis, GENESIS_TIME, BOOT_DEPLOYMENT_FILE, CLIENT_DEPLOYMENT_FILE, get_conf
+from tests.test_bs import current_index, CLIENT_DEPLOYMENT_FILE, get_conf
 from tests.misc import CoreV1ApiClient
-from tests.queries import ES, query_message, query_atx_published
-from elasticsearch_dsl import Search, Q
+from tests.queries import query_atx_published
 from tests.hare.assert_hare import expect_hare
